04_flaskWebService/requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.0.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==1.24.4
oracledb==2.0.1
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
scikit-learn==1.2.1
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.23
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.2
Werkzeug==3.1.4
xgboost==3.1.2
